src/python/sqs-consumer/pyproject.toml
[tool.poetry]
name = "sqs-consumer"
version = "0.1.0"
description = ""
authors = ["pecolynx <pecolynx@gmail.com>"]
readme = "README.md"
packages = [{include = "sqs_consumer"}]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.69"
amazon-sqs-extended-client = "^1.0.0"
pydantic = "^2.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
